+++ package.json
@@ -40,12 +40,14 @@
     "firefox-profile": "^0.3.9",
     "gulp": "~3.5.5",
     "jasmine-core": "^2.2.0",
-    "karma": "^0.12.31",
+    "karma": "^0.13.3",
     "karma-chrome-launcher": "^0.1.7",
     "karma-coverage": "^0.2.7",
     "karma-firefox-launcher": "^0.1.4",
     "karma-jasmine": "^0.3.5",
     "karma-sauce-launcher": "^0.2.14",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^1.7.0",
     "opentok-test-scripts": "^3.0.3",
     "protractor": "3.2.2",
     "q": "^1.2.0",
