+++ tests/unit/index.js
@@ -1,0 +1,7 @@
+require('../../src/js/app.js');
+require('angular-mocks');
+
+require('./controllersSpec.js');
+require('./directivesSpec.js');
+require('./servicesSpec.js');
+require('./subscriberStatsSpec.js');
