+++ tests/unit/login/index.js
@@ -1,0 +1,8 @@
+// var angular = require('angular');
+// require('opentok-angular');
+// require('angular-mocks');
+// angular.module('opentok-meet', ['opentok']);
+// require('../../../src/js/screen/directive.js');
+require('../../../src/js/login/controller.js');
+
+require('./controllerSpec.js');
