+++ tests/unit/screen/index.js
@@ -1,0 +1,9 @@
+// var angular = require('angular');
+// require('opentok-angular');
+// require('angular-mocks');
+// angular.module('opentok-meet', ['opentok']);
+// require('../../../src/js/screen/directive.js');
+require('../../../src/js/screen/controller.js');
+
+require('./controllerSpec.js');
+require('./directiveSpec.js');
